move tests for Replace
